+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Clifton now supports certificate response version 3.
+
 ### Fixed
 - Quote paths in config files. This is required for paths which have spaces in them.
 - Allow `--config-file` to be passed at any level of subcommands.